requirements.txt
fastapi
uvicorn
illama-index
openai
python-dotenv